Extract brush visibility runtime helpers
