+++ CMakeLists.txt
@@ -149,6 +149,7 @@ target_sources(
         src/game_data_brush_editor_export.c
         src/game_data_brush_editor_mutation.c
         src/game_data_brush_trace_queries.c
+        src/game_data_brush_visibility_runtime.c
         src/game_data_brush.c
         src/game_data.c
         src/game_data_actions.c
